Make a new jUnit test for "HelloController"
- Using jUnit.
	- Tested to return "Hello, world!"
