+++ app/src/test/java/com/codesoom/assignment/controllers/HelloControllerWebTest.java
@@ -1,0 +1,27 @@
+package com.codesoom.assignment.controllers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+class HelloControllerWebTest {
+    private final String HELLO = "Hello, world!";
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    void sayHello() throws Exception {
+        mockMvc.perform(get("/"))
+                .andExpect(status().isOk())
+                .andExpect(content().string(HELLO));
+    }
+}
